supercell.kvecs() changed
originally, the list of vectors returned by supercell.kvecs() was sorted using sort_rows(). however, the results of this sorting were unpredictable, in the following way:

construct 2 cubic supercells, one with L=1 and the other with L=8.834763. since both kvecs() returns the vectors in integer coordinates, the two supercells should return the same kvecs (except for last column, magnitudes). however, they did not, because the sorting behavior is unpredictable among vectors of the same magnitude. furthermore, doing this test on different machines results in different sorts as well.

explicitly, i am assuming that the problem is that the result of (a > b) can be unpredictable if a and b are supposed to be the same. to address this, a function is_greater_than() has been added to the math module, which checks first whether a and b are supposed to the same before doing (a > b). a custom insertion sort is then implemented in kvecs() which uses is_greater_than(). the aforementioned test no longer fails in this case